Move registry to daemon/pkg/registry
