+++ daemon/command/config.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/pkg/opts"
+	"github.com/docker/docker/daemon/pkg/registry"
 	dopts "github.com/docker/docker/internal/opts"
-	"github.com/docker/docker/registry"
 	"github.com/spf13/pflag"
 )
