+++ daemon/config/config.go
@@ -14,8 +14,8 @@ import (
 	"dario.cat/mergo"
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/pkg/opts"
+	"github.com/docker/docker/daemon/pkg/registry"
 	dopts "github.com/docker/docker/internal/opts"
-	"github.com/docker/docker/registry"
 	"github.com/moby/moby/api"
 	"github.com/moby/moby/api/types/versions"
 	"github.com/pkg/errors"
