+++ daemon/hosts.go
@@ -16,7 +16,7 @@ import (
 	"github.com/containerd/containerd/v2/core/remotes/docker"
 	hostconfig "github.com/containerd/containerd/v2/core/remotes/docker/config"
 	cerrdefs "github.com/containerd/errdefs"
-	"github.com/docker/docker/registry"
+	"github.com/docker/docker/daemon/pkg/registry"
 	"github.com/pkg/errors"
 )
