+++ daemon/info.go
@@ -19,13 +19,13 @@ import (
 	"github.com/docker/docker/daemon/internal/filedescriptors"
 	"github.com/docker/docker/daemon/internal/metrics"
 	"github.com/docker/docker/daemon/logger"
+	"github.com/docker/docker/daemon/pkg/registry"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/internal/platform"
 	"github.com/docker/docker/pkg/meminfo"
 	"github.com/docker/docker/pkg/parsers/kernel"
 	"github.com/docker/docker/pkg/parsers/operatingsystem"
 	"github.com/docker/docker/pkg/sysinfo"
-	"github.com/docker/docker/registry"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/system"
 	"github.com/opencontainers/selinux/go-selinux"
