+++ daemon/internal/distribution/config.go
@@ -14,8 +14,8 @@ import (
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/layer"
 	refstore "github.com/docker/docker/daemon/internal/refstore"
+	registrypkg "github.com/docker/docker/daemon/pkg/registry"
 	"github.com/docker/docker/pkg/progress"
-	registrypkg "github.com/docker/docker/registry"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/moby/api/types/registry"
 	"github.com/opencontainers/go-digest"
