+++ daemon/internal/distribution/pull.go
@@ -7,7 +7,7 @@ import (
 	"github.com/containerd/log"
 	"github.com/distribution/reference"
 	refstore "github.com/docker/docker/daemon/internal/refstore"
-	"github.com/docker/docker/registry"
+	"github.com/docker/docker/daemon/pkg/registry"
 	"github.com/moby/moby/api/types/events"
 	"github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
