+++ daemon/internal/distribution/push_v2.go
@@ -17,10 +17,10 @@ import (
 	"github.com/docker/docker/daemon/internal/distribution/metadata"
 	"github.com/docker/docker/daemon/internal/distribution/xfer"
 	"github.com/docker/docker/daemon/internal/layer"
+	"github.com/docker/docker/daemon/pkg/registry"
 	"github.com/docker/docker/pkg/ioutils"
 	"github.com/docker/docker/pkg/progress"
 	"github.com/docker/docker/pkg/stringid"
-	"github.com/docker/docker/registry"
 	apitypes "github.com/moby/moby/api/types"
 	"github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
