+++ daemon/internal/distribution/push_v2_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/docker/docker/daemon/internal/distribution/metadata"
 	"github.com/docker/docker/daemon/internal/layer"
 	refstore "github.com/docker/docker/daemon/internal/refstore"
+	registrypkg "github.com/docker/docker/daemon/pkg/registry"
 	"github.com/docker/docker/pkg/progress"
-	registrypkg "github.com/docker/docker/registry"
 	"github.com/moby/moby/api/types/registry"
 	"github.com/opencontainers/go-digest"
 	"gotest.tools/v3/assert"
