+++ daemon/internal/distribution/registry_unit_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/distribution/reference"
-	registrypkg "github.com/docker/docker/registry"
+	registrypkg "github.com/docker/docker/daemon/pkg/registry"
 	"github.com/moby/moby/api/types/registry"
 	"gotest.tools/v3/assert"
 )
