+++ daemon/pkg/plugin/manager.go
@@ -16,10 +16,10 @@ import (
 	"github.com/containerd/containerd/v2/plugins/content/local"
 	"github.com/containerd/log"
 	v2 "github.com/docker/docker/daemon/pkg/plugin/v2"
+	"github.com/docker/docker/daemon/pkg/registry"
 	"github.com/docker/docker/internal/containerfs"
 	"github.com/docker/docker/internal/lazyregexp"
 	"github.com/docker/docker/pkg/authorization"
-	"github.com/docker/docker/registry"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/pubsub"
