+++ daemon/pkg/registry/auth.go
