+++ daemon/pkg/registry/auth_test.go
