+++ daemon/pkg/registry/config.go
