+++ daemon/pkg/registry/config_test.go
