+++ daemon/pkg/registry/errors.go
