+++ daemon/pkg/registry/registry.go
