+++ daemon/pkg/registry/registry_mock_test.go
