+++ daemon/pkg/registry/search.go
