+++ daemon/pkg/registry/search_endpoint_v1.go
