+++ daemon/pkg/registry/search_session.go
