+++ daemon/pkg/registry/search_test.go
