+++ daemon/pkg/registry/service.go
