+++ daemon/pkg/registry/service_v2.go
