+++ daemon/pkg/registry/types.go
