+++ integration/system/login_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
+	registrypkg "github.com/docker/docker/daemon/pkg/registry"
 	"github.com/docker/docker/integration/internal/requirement"
-	registrypkg "github.com/docker/docker/registry"
 	"github.com/moby/moby/api/types/registry"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
